to be merged later

+++ vm/proxy.js
@@ -57,7 +57,7 @@ browser.on('serviceUp', function(service) {
 
   master.on('pull', function(obj) {
   	exec("git pull origin master", {cwd: currentDir}, function() { console.log("MADE A PULL!"); } );
-	vm.kill();
+	if (vm != undefined) vm.kill();
 	vm = launch();
   });
